update the ui states as-in `ui/mod.rs`
